afxdp/tx/dispatch: relocate inline tests to dispatch_tests.rs
Same playbook. 134-LOC inline test block extracted. dispatch.rs: 1,648 → 1,517 LOC.
Note: production code is still over the 1500-LOC threshold; this is a hot-path file
that needs decomposition design (separate PR).
cargo build/test 92/878.
